Rearchitecture for the REANNZ Faucet Project.
These changes were made to allow Faucet to be deployed in the REANNZ
office. We required a vlan capable switch with per-port packet or byte
counters, that we could deploy on hardware in our office.

Requirements:
    - OpenFlow 1.3
    - multiple table
    - correct vlan handling
    - port packet monitoring
    - port status monitoring
    - monitoring disconnections/reconnections of the datapath

This commit contains an implementation for use with the Allied Telesis
x510.

Our initial goal was for quick deployment, so we unfortunately removed a
lot of the functionality from the WAND valve implementation. Sorry!

Removed Functionality:
    - ACLS
    - Support for the hierarchical configuration, configuration for each
    function now occurs at only one place in the config
    - Docker (I assume)

Following are the additions made in this commit:

=================
New Architecture:
=================

valve.py now is a largely controller agnostic implementation of the
switch. It contains functions for specific events (eg. a datapath
connecting to the controller, or a packet being received by the
controller). Each function returns a list of openflow messages (based on
ryu's openflow library) to be sent to the switch, in order (this may
include barrier messages).

The Valve class is now an abstract class that describes the expected
features of a more hardware specific implementation valve. This commit
includes the OVSStatelessValve implementation.

faucet.py is the ryu application that instantiates an implementation of
valve and performs the communication between valve and ryu.

=================
OVSStatelessValve
=================
This is an implementation of valve for use with OpenVSwitch as well as
the Allied Telesis x510. It is stateless in that the controller keeps no
record of mac addresses as they are learned, these are only recorded on
the flow table of the switch. It is also multi table, the table
structure is as follows:
    - Table 0: drops STP BPDUs and LLDP packets, it also drops any
    packet arriving with incorrect vlan tagging, untagged packets
    arriving on ports with native vlans have vlan tags pushed and
    all correctly tagged packets are forwarded to table 1.
    - Table 1: contains entries matching known source mac addresses
    and vlan tags that forward packets to table 2. There is a default
    match sending all other packets (IE those with unknown src mac
    addresses) to the controller and to table 2.
    - Table 2: contains entries matching ethernet destination addresses
    and outputs these to the appropriate ports (stripping vlan tags if
    neccesary) Unmatched packets are sent to table 3.
    - Table 3: matches on vlan and handles flooding. This is in a
    separate table to unicast output rules to allow easy deletion all
    unicast output rules to a specific port without also deleting the
    flood rules for its vlans.

========
gauge.py
========
gauge is the monitoring application. It logs port stats, makes regular
dumps of the datapaths flow table as yaml objects and logs port up/down
and datapath disconnections/re-connections.

=====
tests
=====
A test suite has been added for valve.py. It includes a fake openflow
datapath that performs lookups to verify that the correct openflow
messages have been sent by valve.py in response to certain events to
ensure packets arriving at the datapath are being handled correctly.

=====================
Environment Variables
=====================
New environment variables are used to set up logging and specifying the
location of config files.

=============
configuration
=============
The configuration is now handled within dp.py. This allows config to be
re-used by gauge.py as well as valve.py. There is only configuration for
one datapath per file.
The config now includes an
interfaces and vlans level and names and descriptions for each (Names
are used when logging, descriptions are purely informational). Config
for logging has been added.

============
Native vlans
============
Ports now can have a native vlan as well as tagged vlans on the same
interface. Therefore they are no longer considered tagged or untagged,
they just accept tagged or untagged packets dependant on what vlans are
configured.