Fix pip install
If pbr is not installed, easy_install is required to get latest
version. Setting the PBR_VERSION env variable is breaking the
process. See: https://bugs.launchpad.net/pbr/+bug/1256138

This is a workaround by using a temp METADATA file to force
force faucet's version.

Also the package in pypi is called 'faucet' and not 'ryu-faucet'

Fixes #724